Labeled Graphs

QXGraph uses the SimpleGraph struct from the LightGraphs package to store graph structures. However, some of the algorithms implemented in QXGraph repeatedly modify the graph they work on, either by removing or adding vertices in varying orders, and in turn alter the manner in which vertices in the graph are indexed. This can make it difficult to track where vertices end up in a graph after many modifications are made, which needs to be done if the vertices are used to index different variables in an alternate data structure, such as indices or tensors in a tensor network. To this end, QXGraph defines a LabeledGraph struct which pairs a SimpleGraph with and array of julia symbols which can be used to identify vertices in a graph after modifications have been made.

QXGraph.LabeledGraph — Type

Struct to represent a labeled graph. Unique symbols are created for each vertex if none are provided and remain unaltered during the lifetime of the graph.

Example

julia> g = LabeledGraph()
LabeledGraph({0, 0} undirected simple Int64 graph, Symbol[])

julia> add_vertex!(g, :a_vertex_label)
1-element Array{Symbol,1}:
 :a_vertex_label

julia> g.labels[1]
:a_vertex_label

Example usage

An example of how to use LabeledGraphs is shown below. Note, whenever a modification is made to the graph which re-indexes or re-positions the vertices in the graph, the array of vertex labels in the LabeledGraph is also updated to reflect the reordering. Hence, when a sequence of modifications is made to the graph, we can identify how the vertices of the original graph are now indexed in the new graph by looking at how the corresponding labels are indexed inside the LabeledGraph.

using QXGraph

# Create a LabeledGraph with N vertices
N = 10
G = LabeledGraph(N)

# Display the label assigned to the first and last vertices in the graph.
@show G.labels[1]
@show G.labels[end]

# Remove the first vertex in the graph. To remove a vertex, LightGraphs first swaps the 
# positions of the vertex being removed and the last vertex and then removes the last vertex.
rem_vertex!(G, 1)

# Display the label which is now assigned to the first vertex in the graph.
@show G.labels[1]

LabeledGraph Interface

The interface for the LabeledGraph struct is intended to reflect the interface implemented by the LightGraphs package for the SimpleGraph struct.

QXGraph.get_vertex — Function

get_vertex(G::LabledGraph, v_label)

Return the first vertex whose label matches the argument v_label. If an array of labels is provided, an array of the corresponding vertices is return.

QXGraph.vertices — Function

vertices(G::LabeledGraph)

Return the vertices of a labeled graph.

QXGraph.nv — Function

nv(G::LabeledGraph)

Return the number of vertices in G.

QXGraph.add_vertex! — Function

add_vertex!(G::LabeledGraph, label::Symbol)

Add a new vertex to G and assign the given label to it.

QXGraph.rem_vertex! — Function

rem_vertex!(G::LabeledGraph, v)

Delete the vertex with index or label v from G.

QXGraph.edges — Function

edges(G::LabeledGraph)

Return an iterator of the edges of G.

QXGraph.ne — Function

ne(G::LabeledGraph)

Return the number of edges in G.

QXGraph.add_edge! — Function

add_edge!(G::LabeledGraph, u::Int, v::Int)

Add an edge to G connecting vertices u and v.

QXGraph.has_edge — Function

has_edge(G::LabeledGraph, u::Int, v::Int)

Return true if G has an edge connecting vertices u and v. Return false otherwise.

QXGraph.rem_edge! — Function

rem_edge!(G::LabeledGraph, u::Int, v::Int)

Remove the edge connecting vertices u and v if it exists.

QXGraph.degree — Function

degree(G::LabeledGraph[, v])

Return an array containing the degree of each vertex of G. If v is specified, only return degrees for vertices in v.

QXGraph.all_neighbors — Function

all_neighbors(G::LabeledGraph, v::Int)

Return an array of all neighbors of v in G.

QXGraph.eliminate! — Function

eliminate!(G::LabledGraph, v)

Connect all the neighbors of v together before removing v from G.

Line Graph

QXGraph.line_graph — Function

line_graph(G)

Return a LabeledGraph representing the line graph of the 'G'.

If 'G' is a LabeledGraph then a label for each each vertex of the line graph is created by concatenating the labels of the corresponding vertices in 'G'. Otherwise, labels are created by combining the indices of those vertices in 'G'.

QXGraph.combine_labels — Function

combine_labels(label_A, label_B)

Concatenate 'labelA' and 'labelB' in lexicographical order.