Treewidth Algorithms

One strategy for building contraction plans for tensor networks involves finding a vertex elimination order for the network's line graph. The treewidth of a graph, with respect to a vertex elimination order, is the maximum number of neighbours a vertex has in the graph when it is eliminated according the order. In the context of tensor network contraction, treewidth serves as an indirect measure of the size of the largest intermediate tensor produced while contracting a network according to a the contraction plan built from the elimination order. It is thus also an indirect measure of the computational cost of contracting a tensor network according to the corresponding contraction plan. QXGraph provides functions for finding such elimination orders with minimal treewidth.

Vertex Elimination Orders

A standard algorithm for finding elimination orders, whose treewidth provides a good upper bound for the minimal treewidth of a graph, is known as the QuickBB algorithm. It was first proposed by Vibhav Gogate and Rina Dechter in their 2004 paper "A complete Anytime Algorithm for Treewidth". The paper along with a binary implementation of the algorithm is provided here. QXGraph provides a julia wrapper for their binary which requires a linux OS.

QXGraph.quickbb — Function

quickbb(G::lg.AbstractGraph; 
        time::Integer=0, 
        order::Symbol=:_, 
        verbose::Bool=false )::Tuple{Int, Array{Int, 1}}

Call Gogate's QuickBB binary on the provided graph and return the resulting perfect elimination ordering.

A dictionary containing metadata for the elimination order is also returned. Metadata includes:

:treewidth of the elimination order,

:time taken by quickbb to find the order,

:lowerbound for the treewidth computed by quickbb,

:is_optimal a boolean indicating if the order as optiaml treewidth.

The QuickBB algorithm is described in arXiv:1207.4109v1

Keywords

time::Integer=0: the number of second to run the quickbb binary for.

order::Symbol=:_: the branching order to be used by quickbb (:random or :min_fill).

lb::Bool=false: set if a lowerbound for the treewidth should be computed.

verbose::Bool=false: set to true to print quickbb stdout and stderr output.

proc_id::Integer=0: used to create uniques names of files for different processes.

QXGraph.graph_to_cnf — Function

graph_to_cnf(G::lg.AbstractGraph, filename::String)

Write the provided graph to a cnf file for Gogate's QuickBB binary.

Partial Elimination order

The following functions can be used to create an elimination order for a graph $G$ with a specified clique of $G$ appearing at the end of the order. This is useful for finding contraction plans for tensor networks containing open indices. The algorithm and application of these functions is described further by Shutski et al in this paper.

QXGraph.build_chordal_graph — Function

build_chordal_graph(G::LabeledGraph, π̄::Array{Symbol, 1})

Return a chordal graph built from 'G' using the elimination order 'π̄'.

The returned graph is created from 'G' by iterating over the vertices of 'G', according to the order 'π̄', and for each vertex, connecting all the neighbors that appear later in the order.

QXGraph.restricted_mcs — Function

restricted_mcs(H::LabeledGraph, C::Array{Symbol, 1})

Return an elimination order for 'H' with the vertices in 'C' appearing at the end.

The algorithm is described by Shutski et al in https://arxiv.org/abs/1911.12242

Given an elimination order $\pi$ for a particualr graph $G$, the treewidth of $G$ with respect to the order $\pi$ can be computed using the following function.

QXGraph.find_treewidth_from_order — Function

find_treewidth_from_order(G::LabeledGraph, π̄::Array{Symbol, 1})

Return the treewidth of G with respect to the elimination order π̄.

Treewidth deletion

The problem of finding a select number of vertices in a graph to delete, in order to reduce the treewidth of the graph, is known as the treewidth deletion problem. A method for solving this problem, implemented by the functions below, and it's application to tensor network slicing is discussed by Shutski et al here.

QXGraph.greedy_treewidth_deletion — Function

greedy_treewidth_deletion(G::LabeledGraph, m::Int=4;
                          score_function::Symbol=:degree, 
                          π::Array{Symbol, 1}=[])

Greedily remove vertices from G with respect to minimising the chosen score function. Return the reduced graph and an array of vertices which were removed.

The intermediate elimination orders and corresponding treewidths of the intermediate graphs are also returned if an elimination order for G is provided.

The algorithm is described by Schutski et al in Phys. Rev. A 102, 062614.

Keywords

score_function::Symbol=:degree: function to maximise when selecting vertices to remove. (:degree, :direct_treewidth)

elim_order:Array{Symbol, 1}=Symbol[]: The elimination order for G to be used by direct_treewidth score function.

QXGraph.direct_treewidth_score — Function

direct_treewidth_score(G::LabeledGraph, π::Array{Symbol, 1})

Return an array of integers, one for each vertex in G, indicating the change in treewidth of G, with respect to π, if that vertex is removed.